financial_keepcoders_app/requirements_flask.txt
Flask==2.3.2
joblib==1.4.2
keras==3.5.0
matplotlib==3.9.2
nltk==3.8.1
numpy==1.23.5
pandas==2.2.2
scikeras==0.13.0
seaborn==0.13.2
statsmodels==0.14.2
tensorflow==2.16.2
torch==2.2.2
tqdm==4.66.5
yfinance==0.2.41
TA-Lib==0.4.32
transformers==4.33.0
scikit-learn==1.5.1
python-dotenv==1.0.1
